+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - _Compacted_ `EnumTransformer()`.
 - _Updated_ Osiris scripts.
 
+### REMOVED
+
+- _Cleaned_ `BootstrapServer.lua`.
+
 ## [0.6.9.0] --- 19th July 2020 --- **_WTF is this code?!_**
 
 ### NEW